Initial implementation of functions in order to localize calibrated images and reconstruct the points structure of the scene.
This commit includes
- adds initial functions in order to reconstruct camera poses from calibrated images (+ radially corrected)
  - PointStructureTriangulation: proceed points structures triangulation for a given image ID
  - EuclideanCameraResection: realizes a robust camera resection
  - ReconstructFromTwoCalibratedViews: localizes the two poses using the robust fundamental matrix estimation
  - ExportToPLY: exports the reconstruction into a PLY file
- adds a reconstruction test on synthetic but realistic scene
- adds a incremental reconstruction algorithm into the tracker tool
- changes the tables into vectors in the test_data_sets structure
- adds a function that estimates the RootMeanSquareError in projection.h
- corrects bugs in the classes camera and reconstruction
- changes the variables types from int to size_t in robust_estimation

+++ src/libmv/multiview/robust_estimation.h
@@ -77,13 +77,13 @@ typename Kernel::Model Estimate(const Kernel &kernel,
                                 const Scorer &scorer,
                                 vector<int> *best_inliers = NULL,
                                 double desired_certainty = 0.01) {
-  uint iteration = 0;
+  size_t iteration = 0;
 
-  const int min_samples = Kernel::MINIMUM_SAMPLES;
-  const int total_samples = kernel.NumSamples();
+  const size_t min_samples = Kernel::MINIMUM_SAMPLES;
+  const size_t total_samples = kernel.NumSamples();
 
-  uint max_iterations = 100;
-  const uint really_max_iterations = 1000;
+  size_t max_iterations = 100;
+  const size_t really_max_iterations = 1000;
 
   int best_num_inliers = 0;
   double best_cost = HUGE_VAL;
